requirements.txt
aiogram==2.25.1
Pillow
Flask
python-dotenv
pillow-avif-plugin
pyheif
imageio